--- frontend/components/ProfileDropDown.jsx
@@ -1,108 +1,0 @@
-'use client';
-import { useState, useEffect, useRef } from 'react';
-import Image from 'next/image';
-import profileImage from './profile2.jpg';
-
-const ProfileDropdown = () => {
-  const [isOpen, setIsOpen] = useState(false);
-  const dropdownRef = useRef(null);
-
-  const toggleDropdown = () => {
-    setIsOpen(!isOpen);
-  };
-
-  // Close dropdown when clicking outside
-  useEffect(() => {
-    const handleClickOutside = (event) => {
-      if (dropdownRef.current && !dropdownRef.current.contains(event.target)) {
-        setIsOpen(false);
-      }
-    };
-
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, []);
-
-  return (
-    <div className="relative" ref={dropdownRef}>
-      {/* Profile Icon */}
-      <button
-        onClick={toggleDropdown}
-        className="focus:outline-none"
-      >
-        <div className="relative h-8 w-8 rounded-full overflow-hidden flex items-center justify-center">
-          <Image 
-            src={profileImage} 
-            alt="Profile"
-            width={32}
-            height={32}
-            className="rounded-full object-cover"
-          />
-        </div>
-      </button>
-
-      {/* Dropdown Menu */}
-      {isOpen && (
-        <div className="absolute left-0 mt-2 w-56 rounded-md shadow-lg bg-gray-800 text-gray ring-1 ring-black ring-opacity-5 z-10">
-          <div className="py-1" role="menu" aria-orientation="vertical">
-            <a href="#" className="flex items-center px-4 py-2 text-sm hover:bg-gray-700 transition-colors duration-150" role="menuitem">
-              <span className="mr-3">
-                <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M4 6h16M4 12h16M4 18h16" />
-                </svg>
-              </span>
-              Dashboard
-            </a>
-            <a href="#" className="flex items-center px-4 py-2 text-sm hover:bg-gray-700 transition-colors duration-150" role="menuitem">
-              <span className="mr-3">
-                <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M16 7a4 4 0 11-8 0 4 4 0 018 0zM12 14a7 7 0 00-7 7h14a7 7 0 00-7-7z" />
-                </svg>
-              </span>
-              My profile
-            </a>
-            <div className="border-t border-gray-700 my-1"></div>
-            <a href="#" className="flex items-center px-4 py-2 text-sm hover:bg-gray-700 transition-colors duration-150" role="menuitem">
-              <span className="mr-3">
-                <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 5v2m0 4v2m0 4v2M5 5a2 2 0 00-2 2v3a2 2 0 110 4v3a2 2 0 002 2h14a2 2 0 002-2v-3a2 2 0 110-4V7a2 2 0 00-2-2H5z" />
-                </svg>
-              </span>
-              Redeem code
-            </a>
-            <a href="#" className="flex items-center px-4 py-2 text-sm hover:bg-gray-700 transition-colors duration-150" role="menuitem">
-              <span className="mr-3">
-                <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M12 3v1m0 16v1m9-9h-1M4 12H3m15.364 6.364l-.707-.707M6.343 6.343l-.707-.707m12.728 0l-.707.707M6.343 17.657l-.707.707M16 12a4 4 0 11-8 0 4 4 0 018 0z" />
-                </svg>
-              </span>
-              Light Mode
-            </a>
-            <a href="#" className="flex items-center px-4 py-2 text-sm hover:bg-gray-700 transition-colors duration-150" role="menuitem">
-              <span className="mr-3">
-                <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M10.325 4.317c.426-1.756 2.924-1.756 3.35 0a1.724 1.724 0 002.573 1.066c1.543-.94 3.31.826 2.37 2.37a1.724 1.724 0 001.065 2.572c1.756.426 1.756 2.924 0 3.35a1.724 1.724 0 00-1.066 2.573c.94 1.543-.826 3.31-2.37 2.37a1.724 1.724 0 00-2.572 1.065c-.426 1.756-2.924 1.756-3.35 0a1.724 1.724 0 00-2.573-1.066c-1.543.94-3.31-.826-2.37-2.37a1.724 1.724 0 00-1.065-2.572c-1.756-.426-1.756-2.924 0-3.35a1.724 1.724 0 001.066-2.573c-.94-1.543.826-3.31 2.37-2.37.996.608 2.296.07 2.572-1.065z" />
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M15 12a3 3 0 11-6 0 3 3 0 016 0z" />
-                </svg>
-              </span>
-              Settings
-            </a>
-            <div className="border-t border-gray-700 my-1"></div>
-            <a href="#" className="flex items-center px-4 py-2 text-sm text-red-400 hover:bg-gray-700 transition-colors duration-150" role="menuitem">
-              <span className="mr-3">
-                <svg className="h-5 w-5" fill="none" viewBox="0 0 24 24" stroke="currentColor">
-                  <path strokeLinecap="round" strokeLinejoin="round" strokeWidth={2} d="M17 16l4-4m0 0l-4-4m4 4H7m6 4v1a3 3 0 01-3 3H6a3 3 0 01-3-3V7a3 3 0 013-3h4a3 3 0 013 3v1" />
-                </svg>
-              </span>
-              Log out
-            </a>
-          </div>
-        </div>
-      )}
-    </div>
-  );
-};
-
-export default ProfileDropdown;
