+++ frontend/package.json
@@ -17,6 +17,7 @@
     "@fullcalendar/timegrid": "^6.1.15",
     "@mui/material": "^6.4.3",
     "@mui/x-date-pickers": "^7.27.0",
+    "@phosphor-icons/react": "^2.1.7",
     "axios": "^1.7.9",
     "classnames": "^2.5.1",
     "date-fns": "^4.1.0",
