+++ frontend/src/components/ProfileDropDown.jsx
@@ -1,0 +1,50 @@
+'use client';
+import { useState } from 'react';
+import Image from 'next/image';
+import profileImage from './profile2.jpg';
+import { User, Gear, Question, SignOut } from "@phosphor-icons/react";
+
+const ProfileDropdown = () => {
+  const [toggle, setToggle] = useState(false);
+  const options = [
+    { label: "Profile", icon: <User size={16} className="mr-2" /> },
+    { label: "Settings", icon: <Gear size={16} className="mr-2" /> },
+    { label: "Help", icon: <Question size={16} className="mr-2" /> },
+    { label: "Logout", icon: <SignOut size={16} className="mr-2 text-red-500" /> }
+  ];
+
+  return (
+    <div className="relative">
+      {/* Profile Icon */}
+      <div
+        className="flex items-center gap-2 cursor-pointer p-2 rounded-lg hover:bg-gray-200 transition-all"
+        onClick={() => setToggle(!toggle)}
+      >
+        <Image 
+          src={profileImage} 
+          alt="Profile"
+          width={32}
+          height={32}
+          className="rounded-full object-cover"
+        />
+      </div>
+
+      {/* Dropdown Menu */}
+      {toggle && (
+        <div className="absolute mt-3 w-[130px] bg-white border border-gray-300 rounded-lg shadow-lg z-50"
+          style={{ right: "10px" }}>
+          {options.map((option, index) => (
+            <div
+              key={index}
+              className="px-4 py-3 hover:bg-blue-500 hover:text-white text-black transition-all cursor-pointer flex items-center"
+            >
+              {option.icon} {option.label}
+            </div>
+          ))}
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default ProfileDropdown;
